deep-linking: Enable redirection to narrow
Add isURLNarrow() to link-util for identifying which saved domain owns
the narrow

Fix no-undef error
